refactor(stickiness): move insight to its own folder
